+++ templates/404.blade.php
@@ -1,4 +1,4 @@
-@extends('layouts.base')
+@extends('layouts.app')
 
 @section('content')
   @include('partials.page-header')