+++ templates/page.blade.php
@@ -1,4 +1,4 @@
-@extends('layouts.base')
+@extends('layouts.app')
 
 @section('content')
   @while(have_posts()) @php(the_post())