+++ src/setup.php
@@ -64,7 +64,7 @@ add_action('after_setup_theme', function () {
 
     /**
      * Use main stylesheet for visual editor
-     * @see assets/styles/layouts/_tinymce.scss
+     * @see resources/assets/styles/layouts/_tinymce.scss
      */
     add_editor_style(asset_path('styles/main.css'));
 }, 20);