+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### HEAD
 * Add option to configure build settings ([#1822](https://github.com/roots/sage/pull/1822))
 * Add support for HTML injection ([#1817](https://github.com/roots/sage/pull/1817))
+* Add Tachyons as a CSS framework option
 
 ### 9.0.0-beta.2: January 19th, 2016
 * Fix Browersync ([#1815](https://github.com/roots/sage/pull/1815))